Move compacters to separate file.
