+++ searchlib/src/vespa/searchlib/docstore/CMakeLists.txt
@@ -5,6 +5,7 @@ vespa_add_library(searchlib_docstore OBJECT
     chunk.cpp
     chunkformat.cpp
     chunkformats.cpp
+    compacter.cpp
     data_store_file_chunk_id.cpp
     documentstore.cpp
     document_store_visitor_progress.cpp
